chore: move sharing to new service
