+++ Chefs/Services/Sharing/IShareService.cs
@@ -1,0 +1,9 @@
+namespace Chefs.Services.Sharing;
+
+/// <summary>
+/// Implements content sharing related methods
+/// </summary>
+public interface IShareService
+{
+	Task ShareRecipe(Recipe recipe, IImmutableList<Step> steps, CancellationToken ct);	
+}
